requirements.txt
Jinja2
psycopg2-binary
SQLAlchemy
twine
